+++ service/controller/resource/servicemonitor/service/cluster_autoscaler.go
@@ -88,6 +88,12 @@ func ClusterAutoscaler(cluster metav1.Object, provider string) *promv1.ServiceMo
 			KeyFile:            fmt.Sprintf("/etc/prometheus/secrets/%s/key", key.Secret()),
 			InsecureSkipVerify: true,
 		}
+	} else {
+		serviceMonitor.Spec.Endpoints[0].TLSConfig = &promv1.TLSConfig{
+			CAFile:             key.ControlPlaneCAFile(),
+			InsecureSkipVerify: true,
+		}
+		serviceMonitor.Spec.Endpoints[0].BearerTokenFile = key.ControlPlaneBearerToken()
 	}
 
 	return serviceMonitor